Main renamed the event's own clock and code - `unix` became `currunix`,
`hashcode` became `currhashcode` - stopped owning a market column, and made
every FIX price and quantity exact. This branch was in the middle of turning
`DataType`'s variants into families. The two meet cleanly except where they
name the same thing:

- `rust/tests/fix/schema.rs` keeps main's names and this branch's shapes:
  `DataType::Uuid(UuidType::Uuid)` under `CURRHASHCODE_TAG_NAME`, and the
  Iceberg round trip stays the one this branch rewrote - the digests cross as
  `decimal(20, 0)` through `into_scheme_compat`, the uuid as the spec's own -
  with main's rename applied inside it.
- `rust/tests/fix/dictionary.rs` keeps both histories. Neither pinned hash
  survives a merge of the two, so the literal is what the merged tree
  answers: `2_924_523_166_229_142_018`.
- `docs/hashing.md` takes main's `currhashcode` and this branch's feature
  row, because the `arrow` feature no longer exists to need.
- `docs/extensions/{javascript,python}.md` stay deleted. Main edited them;
  this branch removed the per-language pages and re-homed what only they
  said, so the edit has nowhere to land.
- `.api-inventory.txt` takes main's entries with this branch's paths.

Two sites main added read a sequence by its old variants. `FixMsg::identifier`
walks the `SecurityAltID` group and the CLI's quality check reads a list's
item; both now ask `SequenceType::item`, which is the whole reason the five
layouts share a family.

`docs/assets/fix.json` is main's, and is stale: it is generated from the node
addon, which this branch has not carried through the refactor yet, so it
still spells `u_int64`. It regenerates in the commit that fixes the bindings.

+++ rust/tests/fix/cfb.rs
@@ -9,8 +9,8 @@ use yggdryl::holder::local::Folder;
 
 use yggdryl::holder::Buffer;
 use yggdryl::holder::fs::{File, FileSystem, MemoryFileSystem};
-use yggdryl::{DataType, Error, Field, FixCodec, FixField, FixId, FixRegistry, IOBase};
 use yggdryl::types::SequenceType;
+use yggdryl::{DataType, Error, Field, FixCodec, FixField, FixId, FixRegistry, IOBase};
 
 /// A CBlock in the exact shape a production file has: the same element order,
 /// the same attribute order, the same escaping, the same self-closing forms.
@@ -703,7 +703,9 @@ fn venue_groups_and_their_components_carry_the_membership_and_key_on_the_counter
         );
         assert!(registry.get_field_by_counter(counter).is_some());
     }
-    let DataType::Sequence(SequenceType::List(item)) = roots[0].get_field("vendorentries").unwrap().dtype() else {
+    let DataType::Sequence(SequenceType::List(item)) =
+        roots[0].get_field("vendorentries").unwrap().dtype()
+    else {
         panic!("a list group");
     };
     assert_eq!(branches(item), ["venue"]);
@@ -1523,7 +1525,7 @@ fn folding_a_cblock_into_the_committed_dictionary_refuses_what_it_would_lose() {
         "{message}"
     );
     assert_eq!(seeded.field_by_tag(35).unwrap().dtype(), &DataType::utf8());
-    assert_eq!(seeded.field_by_tag(6).unwrap().dtype(), &DataType::Float64);
+    assert_eq!(seeded.field_by_tag(6).unwrap().dtype(), &DataType::DECIMAL);
 
     assert_eq!(seeded, before, "a refused fold writes nothing");
 }
